requirements.txt
click==8.0.3
jsonschema==4.3.2
paho-mqtt==1.6.1
pymodbus==2.5.3
pyyaml==6.0
tzlocal==4.1
